Updates to handler code
